tweaks

+++ scripts/remove-app.sh
@@ -24,8 +24,6 @@ then
     echo "Error: repo does not exist for ${app_name}"
     exit 1
 fi
-
-
 read -p "Are you sure you want to delete your app permanetly ? 
 (yes/no)  :  " answer
     if [ "$answer" == "no" ];
@@ -35,16 +33,21 @@ read -p "Are you sure you want to delete your app permanetly ?
     elif [ "$answer" == "yes" ];
         then
         echo "deleting ..."
-        for container_id in $(docker compose -f "${YML_FILE}" ps -aq) ;
-            do
-                docker rm -f ${container_id} 2> /dev/null 
-            done
+        if [[ "${YML_FILE}" =~ ^[[:space:]]*$ ]];
+            then 
+                echo "did not find docker-compose.yml ! Dockerfile will be used instead"
+            else
+                for container_id in $(docker compose -f "${YML_FILE}" ps -aq) ;
+                do
+                    docker rm -f ${container_id} 2> /dev/null 
+                done
+            fi
         docker rm -f ${app_name,,}_container 2> /dev/null 
         rm -rf "${APP_PATH}"
         rm -rf "${REPO_PATH}"
         else
-        echo "Invalid input, exiting for safety ..."
-        exit 1
+            echo "Invalid input, exiting for safety ..."
+            exit 1
 
     fi
